new device with name and ports
