+++ cmd/flow.go
@@ -32,20 +32,6 @@ import (
 )
 
 const (
-	// Env vars for port locations
-	PORT_LOCATION_P1 = "${OTG_LOCATION_P1}"
-	PORT_LOCATION_P2 = "${OTG_LOCATION_P2}"
-	// Test port names
-	PORT_NAME_P1 = "p1"
-	PORT_NAME_P2 = "p2"
-	// Env vars for MAC addresses
-	MAC_SRC_P1 = "${OTG_FLOW_SMAC_P1}"
-	MAC_DST_P1 = "${OTG_FLOW_DMAC_P1}"
-	MAC_SRC_P2 = "${OTG_FLOW_SMAC_P2}"
-	MAC_DST_P2 = "${OTG_FLOW_DMAC_P2}"
-	// Default MACs start with "02" to signify locally administered addresses (https://www.rfc-editor.org/rfc/rfc5342#section-2.1)
-	MAC_DEFAULT_SRC = "02:00:00:00:01:aa" // 01 == port 1, aa == otg side (bb == dut side)
-	MAC_DEFAULT_DST = "02:00:00:00:02:aa" // 02 == port 2, aa == otg side (bb == dut side)
 	// Env vars for IPv4 addresses
 	IPV4_SRC = "${OTG_FLOW_SRC_IPV4}"
 	IPV4_DST = "${OTG_FLOW_DST_IPV4}"
@@ -350,13 +336,4 @@ func envSubstOrDefault(e string, d string) string {
 		s = d
 	}
 	return s
-}
-
-func otgConfigHasPort(config gosnappi.Config, name string) bool {
-	for _, p := range config.Ports().Items() {
-		if p.Name() == name {
-			return true
-		}
-	}
-	return false
 }
